feat(source)!: include up to two artist names instead of Various Artists for folder name. If more than two artists but only one DJ then use the DJ's name. #86
This is a semi-breaking change as the naming of .torrent files and output directories will change so some transcoded but not uploaded sources may need to be re-transcoded. Hardly the end of the world but worth noting.

This may also lead to an increase in path length verification failures. If so we can look at making this optional.